fix(ci): remove redundant vision test casts

+++ llama-crab/tests/gemma4_vision.rs
@@ -85,7 +85,7 @@ fn gemma4_vision_question_answering() {
         if let Ok(piece) = llama.model().detokenize(&[tok], false) {
             out.push_str(&piece);
         }
-        let single = LlamaBatch::one(tok, new_n_past + n_generated as i32, 0, true);
+        let single = LlamaBatch::one(tok, new_n_past + n_generated, 0, true);
         llama
             .context()
             .decode(&single)
